Update build dependency & port used w/ default Gulp command
I added imagemin-optipng was a dev-sep because I had to install it to
first run any Gulp tasks. Also, I noticed that when the default Gulp
task was run, it would try to open localhost:4000 instead of
localhost:3011… updating the port from 4000 to 3011 fixed this. Lastly,
I included the minified scripts file in this commit since it updated
after the Gulp task.